backend/requirements.txt
fastapi[standard]==0.116.1
SQLAlchemy==2.0.43
pydantic==2.11.7
python-dotenv==1.1.1
bcrypt==4.0.1
asyncpg==0.30.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
onnxruntime==1.23.2
pillow==12.1.0
pydantic-settings==2.12.0
boto3==1.42.68
python-slugify==8.0.4
opencv-python==4.13.0.92